[FIX] forwardport: fix deduplication of authorship in multi-pr conflict
a45f7260fa36acfdcca1e03c3f388f4f93e96e07 had intended to use the
original authorship information for conflict commit even if there were
multiple commits, as long as there was only one author (/ committer)
for the entire sequence.

Sadly the deduplication was buggy as it took the *authorship date* in
account, which basically ensured commits would never be considered as
having the same authorship outside of tests (where it was possible for
commits to be created at the same second).

Related to #505